requirements.txt
mlflow==2.7.1
wandb==0.15.12
numpy==1.26.1
Pillow==10.1.0
tqdm==4.66.1
seaborn==0.13.0
matplotlib==3.8.0
PyLaTeX==1.4.2
omegaconf==2.3.0
keyboard==0.13.5
albumentations==1.3.1
pycocotools==2.0.7
xmltodict==0.13.0